Renamed files to fit quality criteria
