--- src/pages/admin/TableDashboard.module.css
@@ -1,16 +1,0 @@
-.dashboardTitle {
-    color: #0AA64D;
-    margin-top: 1rem;
-    vertical-align: middle;
-}
-
-.addProduct {
-    color: white;
-}
-
-.productImage {
-    height: auto;
-    width: auto;
-    max-width: 100px;
-    border-radius: 100%;
-}
